separate currency utility functions into separate files
