+++ packages/cardpay-sdk/sdk/currency-utils/types.ts
@@ -1,0 +1,6 @@
+import BigNumber from 'bignumber.js';
+
+/**
+ * Arguments acceptable to a BigNumber constructor, currently equivalent to BigNumber.Value
+ */
+export type BigNumberish = BigNumber.Value;
